Further split student assessment instance page into components
